Rework the patch
Why patch?

My need for this patch can be particular because I could not compile the
source on my machine. The mantainer of Dockerfile do a read of number of
CPU threads and adds one more to use as parallel jobs on compilation,
this was causing errors and interrupting the process. besides that, it
adds support for read the Fulcrum configuration file.